Move tooltip into separate component
+ clarify message
